product/pyproject.toml
[project]
name = "subprime"
version = "0.1.0"
description = "Measuring hidden bias in LLM financial advisors"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Kamal Govindraj", email = "kamal.gs@gmail.com" },
]

dependencies = [
    "pydantic[email]>=2.0",
    "pydantic-ai>=1.0,<2",
    "pydantic-settings>=2.0",
    "anthropic>=0.40",
    "httpx>=0.27",
    "duckdb>=1.2",
    "rich>=13.0",
    "typer>=0.12",
    "scipy>=1.14",
    "numpy>=2.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "jinja2>=3.1",
    # Bumped for CVE-2026-40347 — anything <0.0.26 is vulnerable.
    "python-multipart>=0.0.26",
    "markdown>=3.7",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "playwright>=1.58.0",
    "pillow>=12.2.0",
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-http>=1.27",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "reportlab>=4.2",
    "openpyxl>=3.1",
    "boto3>=1.35",
]

[project.optional-dependencies]
gradio = ["gradio>=5.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "respx>=0.22",
    "ruff>=0.8",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-httpx>=0.30",
    "respx>=0.22",
    "ruff>=0.8",
    "pre-commit>=3.8",
    "mutmut>=3.0",
    "pypdf>=6.0",
]

[project.scripts]
subprime = "subprime.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/subprime"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint.per-file-ignores]
# Legacy modules — imports appear after signal/sys.path setup blocks
# (main.py) or after conditional branches (api.py, gradio_app.py, cli.py).
"apps/web/main.py"                  = ["E402"]
"apps/web/api.py"                   = ["E402", "F841"]
"apps/web/gradio_app.py"            = ["E402"]
"src/subprime/cli.py"               = ["E402"]
# plan_report.py uses semicolon-joined short statements for ReportLab
# column-spec blocks where one line per statement wrecks the table layout.
"src/subprime/core/plan_report.py"  = ["E702", "E741"]
# Tests inline imports after fixture blocks / parametrization markers.
"tests/**"                          = ["E402", "E741"]
